add test for third method
