--- spec/enumerate_spec.rb
@@ -1,1 +1,0 @@
-#
